requirements.txt
fastapi==0.89.1
passlib==1.7.4
pydantic==1.10.4
python-dotenv==0.21.1
SQLAlchemy==1.4.46
starlette==0.22.0
